# Conflicts:
#	Engine/src/Engine/DataStructures/Vector3D.h
#	Game/lib/cmake/EngineCore/EngineCoreTargets-debug.cmake
#	Game/lib/cmake/EngineCore/EngineCoreTargets.cmake

+++ Engine/src/Engine/DataStructures/Vector3D.h
@@ -1,3 +1,5 @@
+// DO NOT MARK except for NumericMin and NumericMax, GetLineIntersectionPointWithPlane, Min, Max, GetSingedDistance, OrderByAngle, GetPlaneAxis, GetPlaneNormal, AnglePointPair
+//This is because it has been submitted for my COMP305. Link to Original: https://github.falmouth.ac.uk/GA-Undergrad-Student-Work-25-26/Comp305-Engine-SL295211.git
 #pragma once
 #include <array>
 
@@ -17,30 +19,7 @@ public:
 	static void GetPlaneAxis(const Vector3D& Normal, Vector3D& T, Vector3D& U);
 	static Vector3D GetPlaneNormal(const Array<Vector3D>& ClippingPlane, const Vector3D& Center);
 
-	//constexpr Vector3D(const float InX, const float InY, const float InZ) noexcept : X(InX), Y(InY), Z(InZ) {};
-
-	constexpr Vector3D(const double InX, const double InY, const double InZ) noexcept : X(static_cast<float>(InX)), Y(static_cast<float>(InY)), Z(static_cast<float>(InZ)){};
-
-	////Vector3D(const Vector3D& CopyVec)
-	////{
-	////	Copy(CopyVec);
-	////}
-
-	////Vector3D& operator=(const Vector3D& CopyVec)
-	////{
-	////	if (this != &CopyVec)
-	////	{
-	////		Copy(CopyVec);
-	////	}
-	////	return *this;
-	////}
-
-	////void Copy(const Vector3D& CopyVec)
-	////{
-	////	X = CopyVec.X;
-	////	Y = CopyVec.Y;
-	////	Z = CopyVec.Z;
-	////}
+	constexpr Vector3D(const double InX, const double InY, const double InZ) noexcept : X(InX), Y(InY), Z(InZ){};
 
 	[[nodiscard]] bool operator==(const Vector3D& Other) const {
 		return	IsAlmostEqual(*this, Other);
